validando dimension de img y decimales

+++ database/migrations/2022_09_25_212629_create_jugadores_table.php
@@ -18,7 +18,7 @@ class CreateJugadoresTable extends Migration
             $table->unsignedBigInteger("IdEquipo");
             $table->unsignedBigInteger("IdCategoria");
             $table->unsignedInteger("IdPersona");
-            $table->integer("EstaturaJugador");
+            $table->float("EstaturaJugador",1,2);
             $table->string("FotoCarnet");
             $table->string("PosicionJugador");
             $table->integer("NumeroCamiseta");
